qcacld-3.0: Fix the 64-bit division linking error for msmcobalt_32
The 64-bit division in 32-bit architecture call 64-bit-by-64-bit
division routines "__aeabi_uldivmod", which is not defined for
the 32-bit architecture. Typecast the variables to proper data
type before division operation to avoid the "undefined symbol:
 __aeabi_uldivmod" linking error during wlan driver load on
32-bit target.

CRs-Fixed: 1091920

+++ core/hdd/src/wlan_hdd_cfg80211.c
@@ -10739,7 +10739,7 @@ struct cfg80211_bss *wlan_hdd_cfg80211_inform_bss_frame(hdd_adapter_t *pAdapter,
 	 * send age in units of 1/10 ms.
 	 */
 	qie_age->age =
-		(qdf_mc_timer_get_system_time() - bss_desc->received_time)/10;
+		(uint32_t)(qdf_mc_timer_get_system_time() - bss_desc->received_time)/10;
 	qie_age->tsf_delta = bss_desc->tsf_delta;
 	memcpy(&qie_age->beacon_tsf, bss_desc->timeStamp,
 	       sizeof(qie_age->beacon_tsf));
